Add a resource def to import existing instance
Commit replicates the basic configuration of an already existing VPS in
order to allow the import of the respective instance into TF's state.
The point of this is to ensure that the server remains in place as is
(with a plan no longer available in its location and an IP address
already referenced outside of the cloud), with a view at upgrading its
outdated OS via Terraform - subsequent config would be done via Ansible.