Add a config for the bastion's firewall
The pf.conf should protect the bastion for abuses on its open SSH port;
this is just one layer in the defense-in-depth measures that will need
to be taken. The .gitignore file had to be modified as well - the
current approach for ignoring files/dirs might need to be revisited.