Read sshfp via data.read_file instead of file()
This fixes a dependency issue caused by the fact that file() requires
that its target exist when a TF plan/apply is started, but the sshfp
files would require an ongoing apply to be created/updated.

Even this approach might need replacing, but it will certainly need that
the TODO comments found in the commit be resolved.